component sockets and swapping
this is a gross summarisation of a painful amount of work to get all components properly attachable, detachable, and dynamically integrated into all ship systems.
you can now swap weapons. you can now add more capacitors and coolers.
this is the framework for a very fun time.
it hurt. it's awesome.