item spawning testing
i think the way to go with item spawning is have none of the cliens *need* to reference an item. they just request spawns from the server.

also gonna have to rework the inventory systems with this to store purely data about the items and not items themselves. will make things a lot easier